+++ relevanceai/clusterer/clusterer.py
@@ -385,7 +385,6 @@ class ClusterOps(BatchAPIClient):
     @track
     def aggregate(
         self,
-        dataset: Optional[Union[str, Dataset]] = None,
         vector_fields: List[str] = None,
         metrics: list = [],
         sort: list = [],
@@ -395,6 +394,7 @@ class ClusterOps(BatchAPIClient):
         page: int = 1,
         asc: bool = False,
         flatten: bool = True,
+        dataset: Optional[Union[str, Dataset]] = None,
     ):
         """
         Takes an aggregation query and gets the aggregate of each cluster in a collection. This helps you interpret each cluster and what is in them.
@@ -1319,7 +1319,7 @@ class ClusterOps(BatchAPIClient):
             ]
         return new_documents
 
-    def _set_cluster_labels_across_documents(self, cluster_labels, documents):
+    def _get_cluster_field_name(self):
         if isinstance(self.vector_fields, list):
             set_cluster_field = (
                 f"{self.cluster_field}.{'.'.join(self.vector_fields)}.{self.alias}"
@@ -1328,6 +1328,10 @@ class ClusterOps(BatchAPIClient):
             set_cluster_field = (
                 f"{self.cluster_field}.{self.vector_fields}.{self.alias}"
             )
+        return set_cluster_field
+
+    def _set_cluster_labels_across_documents(self, cluster_labels, documents):
+        set_cluster_field = self._get_cluster_field_name()
         self.set_field_across_documents(set_cluster_field, cluster_labels, documents)
 
     def _label_cluster(self, label: Union[int, str]):
@@ -1527,3 +1531,42 @@ class ClusterOps(BatchAPIClient):
                 }
 
         return stats
+
+    def report(self):
+        """
+        Get a report on your clusters.
+
+        Example
+        ---------
+        .. code-block::
+
+            from relevanceai.datasets import mock_documents
+            docs = mock_documents(10)
+            df = client.Dataset('sample')
+            df.upsert_documents(docs)
+            cluster_ops = df.auto_cluster('kmeans-2', ['sample_1_vector_'])
+            cluster_ops.report()
+
+        """
+        if isinstance(self.vector_fields, list) and len(self.vector_fields) > 1:
+            raise ValueError(
+                "We currently do not support more than 1 vector field when reporting."
+            )
+        from relevanceai.cluster_report import ClusterReport
+
+        # X is all the vectors
+        cluster_field_name = self._get_cluster_field_name()
+        all_docs = self._get_all_documents(
+            self.dataset_id, select_fields=self.vector_fields + [cluster_field_name]
+        )
+        cluster_labels = self.get_field_across_documents(cluster_field_name, all_docs)
+        self.number_of_clusters = len(set(cluster_labels))
+        self._report = ClusterReport(
+            self.get_field_across_documents(self.vector_fields[0], all_docs),
+            cluster_labels=self.get_field_across_documents(
+                cluster_field_name, all_docs
+            ),
+            model=self.model,
+            num_clusters=self.number_of_clusters,
+        )
+        return self._report.get_cluster_internal_report()